Remake of architecture
