+++ .env.example
@@ -1,0 +1,4 @@
+BOT_TOKEN = 111
+RAPID_API_HOST = 222
+RAPID_API_KEY = 333
+DATABASE_NAME = 444
