+++ handlers/__init__.py
@@ -1,0 +1,2 @@
+from . import custom_handlers
+from . import default_handlers
