+++ handlers/default_handlers/__init__.py
@@ -1,0 +1,3 @@
+from . import start
+from . import help
+from . import echo
