+++ keyboards/__init__.py
@@ -1,0 +1,2 @@
+from . import reply
+from . import inline
