+++ utils/set_bot_commands.py
@@ -1,0 +1,8 @@
+from telebot.types import BotCommand
+from config_data.config import DEFAULT_COMMANDS
+
+
+def set_default_commands(bot):
+    bot.set_my_commands(
+        [BotCommand(*i) for i in DEFAULT_COMMANDS]
+    )
